// generated/src/app/api-models/stage-auth-prov.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthProv {
  batchId: string;
  transactionId: number;
  seqNo: number;
  provCode: string;
  provId: string;
  provTaxId: string;
  provNatId: string;
  provLastName: string;
  provFirstName: string;
  provAddress: string;
  provCity: string;
  provState: string;
  provZip: string;
  outOfNetStatus: string;
  provIdQualifier: string;
  provEntityIdCode: string;
  provTaxonomyCode: string;
  insertDatetime: Date;
}